fuuck

+++ src/world/get_sector.c
@@ -6,7 +6,7 @@
 /*   By: mfischer <mfischer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/10 23:54:27 by mfischer          #+#    #+#             */
-/*   Updated: 2019/05/10 23:58:09 by mfischer         ###   ########.fr       */
+/*   Updated: 2019/05/11 16:45:29 by mfischer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ t_sector *get_sector(int id, t_world *world)
 	int i;
 
 	i = -1;
-	while (++i < world->sectornum)
+	while (++i < (int)world->sectornum)
 		if (world->sectors[i].id == id)
 			return (&world->sectors[i]);
 	return (NULL);
